Add files via upload
My research focuses on the problem of identifying how cracks develop in homogeneous materials under stress, including the material's capacity to bear loads over its breaking point. This entails taking into account a number of variables, such as the material's behavior (brittle or ductile), its characteristics, the specimen's size and shape, how it is loaded, and any potential imperfections or flaws. The maximum load capacity of these materials is usually ascertained by experiments and theoretical models. Precise prediction is challenging, nevertheless, due to the intricacy of fracture behavior, especially in quasi-brittle materials.

I used a dataset that gives precise information about the location and dimensions of a hole inside a material for the project/problem statement. These particulars are crucial since they affect the material's ability to support loads and maintain its structural integrity. I utilized the output frequencies found in the dataset as the inputs for a machine learning model. Based on these frequencies, this model seeks to forecast the features of the hole, including its precise location and dimensions. I contributed to the assessment of the precision and potency of my method by contrasting the model's predictions with the actual measurements. 

The usage of frequencies in my research could have anything to do with Acoustic Emission (AE) testing, a method that analyzes sound waves and energy released by structural changes in a material. AE is frequently used to track seismic activity, examine material deterioration, and even identify or anticipate failures. Although it's unclear whether the frequencies used in my research come directly from AE testing, they have a similar function in that they provide information on the material's behavior under stress, which is crucial for determining its integrity and forecasting problems.